Close #26 - Resolve DDB table names at service startup time
Stage value is configured at Ninja configuration file for different
Ninja mode. Stage value is then passed through the EnvironmentContext
to the repositories.
