+++ instand-service/src/main/java/conf/Constants.java
@@ -1,0 +1,9 @@
+package conf;
+
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
+public class Constants {
+
+    public static final String STAGE_KEY_NAME = "stage";
+}
